[udev][usb-mass-storage] Fix queued iotxn list corruption problem
We had been delaying pulling iotxns off queued_iotxns list
until they were completed so the IOCTL_DEVICE_SYNC implementation
will see them. But it is unsafe to do this since completing an iotxn
will likely result in that iotxn being added to another list.
So now we pop the iotxn off the queued_iotxns before processing
and keep it in a new curr_txn field so the IOCTL_DEVICE_SYNC can access it.

Change-Id: I6be2ddc3a216337fd8a655e51edde26a61227bd4